server/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
black==23.3.0
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cors==1.0.1
cryptography==40.0.2
cycler==0.11.0
dataclasses-json==0.5.7
Deprecated==1.2.13
filelock==3.12.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
fonttools==4.39.3
frozenlist==1.3.3
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langchain==0.0.157
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
plotly==5.14.1
pycparser==2.21
pydantic==1.10.7
PyGithub==1.58.1
PyJWT==2.6.0
PyNaCl==1.5.0
pyodbc==4.0.39
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.4
regex==2023.5.5
requests==2.29.0
requests-file==1.5.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.12
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
tldextract==3.4.1
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.3
wrapt==1.15.0
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
